+++ app/src/main/java/de/baumann/browser/helper/Activity_settings_search.java
@@ -23,6 +23,7 @@ import android.os.Bundle;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.app.AppCompatDelegate;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
@@ -41,6 +42,8 @@ public class Activity_settings_search extends AppCompatActivity {
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         setTitle(R.string.action_searchChoose);
+
+        AppCompatDelegate.setCompatVectorFromResourcesEnabled(true);
 
         android.support.v7.app.ActionBar actionBar = getSupportActionBar();
         if(actionBar != null) {